# contracts/openapi/openapi.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Evidrilo API",
    "version": "1.0.0",
    "description": "Public contract for the versioned Evidrilo API routes."
  },
  "paths": {
...
    "/v2/project-ai/general-chat": {
      "post": {
        "operationId": "post_v2_project_ai_general_chat",
        "tags": [
          "v2"
        ],
        "security": [
          {
            "SupabaseBearer": []
          }
        ],
        "responses": {
          "2XX": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/project-ai-general-chat-result.v2"
                }
              }
            }
          },
          "default": {
            "description": "API error response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/http-errors.v1"
                }
              }
            }
          }
        },
        "x-evidrilo-test": {
          "file": "platform/api.Tests/ProjectAiScaffoldFlowTests.cs",
          "anchor": "/v2/project-ai/general-chat"
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/project-ai-general-chat-request.v2"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "description": "Stable request key used to prevent duplicate General chat dispatches.",
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 128,
              "pattern": "^[A-Za-z0-9_-]+$"
            }
          },
          {
            "name": "X-Evidrilo-General-Chat-Consent",
            "in": "header",
            "required": true,
            "description": "Must equal general-chat.v1 to confirm the current General chat notice for this request.",
            "schema": {
              "type": "string",
              "const": "general-chat.v1"
            }
          }
        ]
      },
      "parameters": []
    }
  },
  "components": {
    "securitySchemes": {
      "SupabaseBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "RevenueCatWebhookAuthorization": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      },
      "RevenueCatWebhookSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-RevenueCat-Webhook-Signature"
      }
    },
    "schemas": {
...
      "http-errors.v1": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "https://evidrilo.dev/contracts/http-errors.v1.json",
        "title": "Evidrilo HTTP error response v1",
        "type": "object",
        "additionalProperties": false,
        "required": [
          "schema",
          "version",
          "code",
          "message",
          "requestId"
        ],
        "properties": {
          "schema": {
            "const": "evidrilo.http-error"
          },
          "version": {
            "const": "1"
          },
          "code": {
            "type": "string",
            "pattern": "^[A-Z][A-Z0-9_]{2,63}$"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 240
          },
          "requestId": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128,
            "pattern": "^[A-Za-z0-9_-]+$"
          },
          "creditCost": {
            "description": "Present only when verified provider usage was charged even though the requested response was withheld.",
            "type": "integer",
            "minimum": 1,
            "maximum": 200
          }
        }
      },
...
      "project-ai-general-chat-request.v2": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "https://evidrilo.dev/contracts/project-ai-general-chat-request.v2.json",
        "title": "Evidrilo isolated Project AI General chat request v2",
        "description": "Standalone message only. Project identifiers, project context, and caller-supplied history are not accepted.",
        "type": "object",
        "additionalProperties": false,
        "required": [
          "schema",
          "version",
          "installationId",
          "locale",
          "message"
        ],
        "properties": {
          "schema": {
            "const": "evidrilo.project-ai-general-chat"
          },
          "version": {
            "const": "2"
          },
          "installationId": {
            "type": "string",
            "format": "uuid"
          },
          "locale": {
            "type": "string",
            "maxLength": 32,
            "pattern": "^[A-Za-z]{2,8}(?:-[A-Za-z0-9]{1,8})*$"
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 4000
          }
        }
      },
      "project-ai-general-chat-result.v2": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "$id": "https://evidrilo.dev/contracts/project-ai-general-chat-result.v2.json",
        "title": "Evidrilo isolated Project AI General chat result v2",
        "description": "Structured answer, standalone suggested follow-up prompts, and token-priced credit cost. The message, answer, and prompts are not part of activity history.",
        "type": "object",
        "additionalProperties": false,
        "required": [
          "schema",
          "version",
          "mode",
          "status",
          "answer",
          "recommendedNextPrompts",
          "requestId",
          "creditCost"
        ],
        "properties": {
          "schema": {
            "const": "evidrilo.project-ai-general-chat"
          },
          "version": {
            "const": "2"
          },
          "mode": {
            "const": "GENERAL"
          },
          "status": {
            "const": "success"
          },
          "answer": {
            "type": "string",
            "minLength": 1,
            "maxLength": 8000
          },
          "recommendedNextPrompts": {
            "type": "array",
            "minItems": 1,
            "maxItems": 3,
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 240
            }
          },
          "requestId": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128,
            "pattern": "^[A-Za-z0-9_-]+$"
          },
          "creditCost": {
            "type": "integer",
            "minimum": 1,
            "maximum": 200
          }
        }
      },
...
    }
  }
}